Made reflection cache bootstrap editor only.
