--- Utility/Reflection.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6cbad53ccccbbb845b8826d4e1c9d3b6
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
